CompileViews added
